+++ src/env/browser/fs.ts
@@ -1,0 +1,3 @@
+export function realpath(path: string): Promise<string> {
+	return Promise.resolve(path);
+}